feat(worldgen): move shop assignment to assignShopsToBuildings in shops_core.js and wire it in town_gen.js
